move quota route to ent
